fix(startup error): update @testing-library/dom to update pretty-format
fixes: #519

BREAKING CHANGE: This change will result in a compile error for any `MatcherFunction` that uses `HTMLElement` properties or functions.
`MatcherFunction` in `@testing-library/dom` now receives a parameter of type `Element`,
where it previously was `HTMLElement` - so users will have to test/cast to `HTMLElement`
if they need `HTMLElement` properties or methods in their matcher function.