fix: remove `waitForAsync` (zoneless support)
In this commit, we remove `waitForAsync` as it is no longer necessary. This rule has also been
disabled internally at Google and is no longer in use. Its removal enables zoneless support,
since `waitForAsync` depends on zone.js.

We also remove `compileComponents()`, as it is only required when `setClassMetadataAsync`
is called—typically representing AOT pre-compiled components outside of TestBed.